clean up tests.
