+++ src/test/java/io/pivotal/accounts/service/AccountServiceTest.java
@@ -20,9 +20,11 @@ import io.pivotal.accounts.repository.AccountRepository;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 /**
@@ -31,21 +33,15 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
  * @author David Ferreira Pinto
  *
  */
+@RunWith(MockitoJUnitRunner.class)
 public class AccountServiceTest {
-	MockMvc mockMvc;
 
 	@InjectMocks
 	AccountService service;
 	
 	@Mock
 	AccountRepository repo;
-	
-	@Before
-	public void setup() {
-		MockitoAnnotations.initMocks(this);
 
-	    this.mockMvc = MockMvcBuilders.standaloneSetup(service).build();
-	}
 
 	/**
 	 * test retrieval of account by integer.
